+++ examples/read_write_both.c
@@ -6,13 +6,13 @@
 #include <stdio.h>
 #include <string.h>
 
-#define Read_INTERFACE method(size_t, read, void *self, char *dest, size_t bytes_to_read)
+#define Read_IFACE method(size_t, read, void *self, char *dest, size_t bytes_to_read)
 interface(Read);
 
-#define Write_INTERFACE method(size_t, write, void *self, const char *src, size_t bytes_to_write)
+#define Write_IFACE method(size_t, write, void *self, const char *src, size_t bytes_to_write)
 interface(Write);
 
-#define ReadWrite_INTERFACE
+#define ReadWrite_IFACE
 #define ReadWrite_EXTENDS (Read, Write)
 interface(ReadWrite);