+++ examples/state.c
@@ -5,7 +5,7 @@
 #include <stdio.h>
 
 // clang-format off
-#define State_INTERFACE                 \
+#define State_IFACE                     \
     method( int, get, void *self)       \
     method(void, set, void *self, int x)
 // clang-format on